Move BinaryComparisonOperation out of the Processor
