+++ x-pack/plugin/esql-core/src/main/java/org/elasticsearch/xpack/esql/core/expression/predicate/operator/comparison/BinaryComparisonProcessor.java
@@ -9,55 +9,10 @@ package org.elasticsearch.xpack.esql.core.expression.predicate.operator.comparis
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.xpack.esql.core.expression.gen.processor.FunctionalEnumBinaryProcessor;
 import org.elasticsearch.xpack.esql.core.expression.gen.processor.Processor;
-import org.elasticsearch.xpack.esql.core.expression.predicate.PredicateBiFunction;
-import org.elasticsearch.xpack.esql.core.expression.predicate.operator.comparison.BinaryComparisonProcessor.BinaryComparisonOperation;
 
 import java.io.IOException;
-import java.util.function.BiFunction;
 
 public class BinaryComparisonProcessor extends FunctionalEnumBinaryProcessor<Object, Object, Boolean, BinaryComparisonOperation> {
-
-    public enum BinaryComparisonOperation implements PredicateBiFunction<Object, Object, Boolean> {
-
-        EQ(Comparisons::eq, "=="),
-        NULLEQ(Comparisons::nulleq, "<=>"),
-        NEQ(Comparisons::neq, "!="),
-        GT(Comparisons::gt, ">"),
-        GTE(Comparisons::gte, ">="),
-        LT(Comparisons::lt, "<"),
-        LTE(Comparisons::lte, "<=");
-
-        private final BiFunction<Object, Object, Boolean> process;
-        private final String symbol;
-
-        BinaryComparisonOperation(BiFunction<Object, Object, Boolean> process, String symbol) {
-            this.process = process;
-            this.symbol = symbol;
-        }
-
-        @Override
-        public String symbol() {
-            return symbol;
-        }
-
-        @Override
-        public Boolean apply(Object left, Object right) {
-            if (this != NULLEQ && (left == null || right == null)) {
-                return null;
-            }
-            return doApply(left, right);
-        }
-
-        @Override
-        public final Boolean doApply(Object left, Object right) {
-            return process.apply(left, right);
-        }
-
-        @Override
-        public String toString() {
-            return symbol;
-        }
-    }
 
     public static final String NAME = "cb";
